requirements.txt
fastapi==0.100.0
uvicorn==0.23.1
pydantic==2.0.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
alembic==1.11.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
redis==4.6.0
kafka-python==2.0.2
httpx==0.24.1
pytest==7.4.0
pydantic-settings==2.0.2
python-dotenv==1.0.0
tenacity==8.2.2
asyncpg==0.28.0
prometheus-fastapi-instrumentator==6.1.0
kubernetes==27.2.0
gunicorn==21.2.0
structlog==23.1.0
ujson==5.8.0 